Cluster re-ordering Moved cluster re-ordering to cluster.py Clusters are rearranged by default but can be turned off if desired minor Python3 compliance + PEP8 cleaning
