+++ pycroscopy/processing/cluster.py
@@ -6,6 +6,8 @@ Created on Tue Jan 05 07:55:56 2016
 """
 import numpy as np
 import sklearn.cluster as cls
+from scipy.cluster.hierarchy import linkage
+from scipy.spatial.distance import pdist
 
 from ..io.hdf_utils import checkIfMain
 from ..io.hdf_utils import getH5DsetRefs, checkAndLinkAncillary
@@ -34,8 +36,8 @@ class Cluster(object):
         *args and **kwargs : arguments to be passed to the estimator
         """
 
-        allowed_methods = ['AgglomerativeClustering','Birch','KMeans',
-                           'MiniBatchKMeans','SpectralClustering']
+        allowed_methods = ['AgglomerativeClustering', 'Birch', 'KMeans',
+                           'MiniBatchKMeans', 'SpectralClustering']
         
         # check if h5_main is a valid object - is it a hub?
         if not checkIfMain(h5_main):
@@ -61,10 +63,15 @@ class Cluster(object):
         self.data_transform_func, self.data_is_complex, self.data_is_compound, \
         self.data_n_features, self.data_n_samples, self.data_type_mult = retval
 
-
-    def doCluster(self):
+    def do_cluster(self, rearrange_clusters=True):
         """
-        Clusters the hdf5 dataset, calculates mean response for each cluster, and writes the labels and mean response back to the h5 file
+        Clusters the hdf5 dataset, calculates mean response for each cluster, and writes the labels and mean response
+        back to the h5 file
+
+        Parameters
+        ----------
+        rearrange_clusters : (Optional) Boolean. Default = True
+            Whether or not the clusters should be re-ordered by relative distances between the mean response
 
         Returns
         --------
@@ -72,8 +79,11 @@ class Cluster(object):
             Reference to the group that contains the clustering results
         """
         self._fit()
-        mean_response = self._getMeanResponse(self.results.labels_)
-        return self._writeToHDF5(self.results.labels_, mean_response)
+        new_mean_response = self._get_mean_response(self.results.labels_)
+        new_labels = self.results.labels_
+        if rearrange_clusters:
+            new_labels, new_mean_response = reorder_clusters(self.results.labels_, new_mean_response)
+        return self._write_to_hdf5(new_labels, new_mean_response)
 
     def _fit(self):
         """
@@ -85,9 +95,8 @@ class Cluster(object):
         """
         # perform fit on the real dataset
         self.results = self.estimator.fit(self.data_transform_func(self.h5_main[self.data_slice]))
-        
 
-    def _getMeanResponse(self, labels):
+    def _get_mean_response(self, labels):
         """
         Gets the mean response for each cluster
 
@@ -103,18 +112,18 @@ class Cluster(object):
         """
         num_clusts = len(np.unique(labels))
         mean_resp = np.zeros(shape=(num_clusts, self.num_comps), dtype=self.h5_main.dtype)
-        for clust_ind in xrange(num_clusts):
+        for clust_ind in range(num_clusts):
             # get all pixels with this label
             targ_pos = np.where(labels == clust_ind)[0]
             # slice to get the responses for all these pixels, ensure that it's 2d
             data_chunk = np.atleast_2d(self.h5_main[targ_pos, self.data_slice[1]])
-            #transform to real from whatever type it was
+            # transform to real from whatever type it was
             avg_data = np.mean(self.data_transform_func(data_chunk), axis=0, keepdims=True)
             # transform back to the source data type and insert into the mean response
             mean_resp[clust_ind] = transformToTargetType(avg_data, self.h5_main.dtype)
         return mean_resp
 
-    def _writeToHDF5(self, labels, mean_response):
+    def _write_to_hdf5(self, labels, mean_response):
         """
         Writes the labels and mean response to the h5 file
 
@@ -183,3 +192,47 @@ class Cluster(object):
 
         # return the h5 group object
         return h5_labels.parent
+
+
+def reorder_clusters(labels, mean_response):
+    """
+    Reorders clusters by the distances between the clusters
+
+    Parameters
+    ----------
+    labels : 1D unsigned int numpy array
+        Labels for the clusters
+    mean_response : 2D numpy array
+        Mean response of each cluster arranged as [cluster , features]
+
+    Returns
+    -------
+    new_labels : 1D unsigned int numpy array
+        Labels for the clusters arranged by distances
+    new_mean_response : 2D numpy array
+        Mean response of each cluster arranged as [cluster , features]
+    """
+
+    num_clusters = mean_response.shape[0]
+    # Get the distance between cluster means
+    distance_mat = pdist(mean_response)
+    # get hierarchical pairings of clusters
+    linkage_pairing = linkage(distance_mat, 'weighted')
+
+    # get the new order - this has been checked to be OK
+    new_cluster_order = []
+    for row in range(linkage_pairing.shape[0]):
+        for col in range(2):
+            if linkage_pairing[row, col] < num_clusters:
+                new_cluster_order.append(int(linkage_pairing[row, col]))
+
+    # Now that we know the order, rearrange the clusters and labels:
+    new_labels = np.zeros(shape=labels.shape, dtype=labels.dtype)
+    new_mean_response = np.zeros(shape=mean_response.shape, dtype=mean_response.dtype)
+
+    # Reorder clusters
+    for old_clust_ind, new_clust_ind in enumerate(new_cluster_order):
+        new_labels[np.where(labels == new_clust_ind)[0]] = old_clust_ind
+        new_mean_response[old_clust_ind] = mean_response[new_clust_ind]
+
+    return new_labels, new_mean_response
